Convert ExplorerPanel back to using XAML
Remove Project Refresh button (no longer required)
