+++ app/Workspace/Celbridge.Explorer/Views/ExplorerPanel.xaml
@@ -1,0 +1,36 @@
+﻿<UserControl
+    x:Class="Celbridge.Explorer.Views.ExplorerPanel"
+    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+    xmlns:local="using:Celbridge.Explorer.Views"
+    xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+    mc:Ignorable="d"
+    DataContext="{x:Bind ViewModel}">
+
+  <Grid VerticalAlignment="Stretch">
+    <Grid.RowDefinitions>
+      <RowDefinition Height="40"/>
+      <RowDefinition Height="*"/>
+    </Grid.RowDefinitions>
+
+    <!-- Title Bar -->
+    <Grid Grid.Row="0"
+          BorderBrush="{ThemeResource PanelBorderBrush}"
+          BorderThickness="0,0,0,1">
+      <Grid.ColumnDefinitions>
+        <ColumnDefinition Width="Auto"/>
+        <ColumnDefinition Width="*"/>
+      </Grid.ColumnDefinitions>
+
+      <TextBlock Grid.Column="0"
+                 Text="{x:Bind ViewModel.TitleText, Mode=OneWay}"
+                 Margin="48,0,0,0"
+                 VerticalAlignment="Center"/>
+    </Grid>
+
+    <!-- Resource Tree View -->
+    <local:ResourceTreeView x:Name="ResourceTreeView"
+                            Grid.Row="1"/>
+  </Grid>
+</UserControl>
